Moved spamd check to a custom file and cached the result
