+++ app/models.py
@@ -3,7 +3,7 @@ import os
 import random
 import uuid
 from email.utils import formataddr
-from typing import List, Tuple, Optional, Dict
+from typing import List, Tuple, Optional
 
 import arrow
 import sqlalchemy as sa
@@ -235,75 +235,6 @@ class AuditLogActionEnum(EnumE):
     disable_2fa = 5
     logged_as_user = 6
     extend_subscription = 7
-
-
-class Phase(EnumE):
-    unknown = 0
-    forward = 1
-    reply = 2
-
-
-class DmarcCheckResult(EnumE):
-    allow = 0
-    soft_fail = 1
-    quarantine = 2
-    reject = 3
-    not_available = 4
-    bad_policy = 5
-
-    @staticmethod
-    def get_string_dict():
-        return {
-            "DMARC_POLICY_ALLOW": DmarcCheckResult.allow,
-            "DMARC_POLICY_SOFTFAIL": DmarcCheckResult.soft_fail,
-            "DMARC_POLICY_QUARANTINE": DmarcCheckResult.quarantine,
-            "DMARC_POLICY_REJECT": DmarcCheckResult.reject,
-            "DMARC_NA": DmarcCheckResult.not_available,
-            "DMARC_BAD_POLICY": DmarcCheckResult.bad_policy,
-        }
-
-
-class SPFCheckResult(EnumE):
-    allow = 0
-    fail = 1
-    soft_fail = 1
-    neutral = 2
-    temp_error = 3
-    not_available = 4
-    perm_error = 5
-
-    @staticmethod
-    def get_string_dict():
-        return {
-            "R_SPF_ALLOW": SPFCheckResult.allow,
-            "R_SPF_FAIL": SPFCheckResult.fail,
-            "R_SPF_SOFTFAIL": SPFCheckResult.soft_fail,
-            "R_SPF_NEUTRAL": SPFCheckResult.neutral,
-            "R_SPF_DNSFAIL": SPFCheckResult.temp_error,
-            "R_SPF_NA": SPFCheckResult.not_available,
-            "R_SPF_PERMFAIL": SPFCheckResult.perm_error,
-        }
-
-
-class SpamdResult:
-    def __init__(self, phase: Phase = Phase.unknown):
-        self.phase: Phase = phase
-        self.dmarc: DmarcCheckResult = DmarcCheckResult.not_available
-        self.spf: SPFCheckResult = SPFCheckResult.not_available
-
-    def set_dmarc_result(self, dmarc_result: DmarcCheckResult):
-        self.dmarc = dmarc_result
-
-    def set_spf_result(self, spf_result: SPFCheckResult):
-        self.spf = spf_result
-
-    def event_data(self) -> Dict:
-        return {
-            "header": "present",
-            "dmarc": self.dmarc,
-            "spf": self.spf,
-            "phase": self.phase,
-        }
 
 
 class Hibp(Base, ModelMixin):
